+++ emr-serverless-spark-20230808/alibabacloud_emr_serverless_spark20230808/models/_create_kyuubi_token_request.py
@@ -13,12 +13,14 @@ class CreateKyuubiTokenRequest(DaraModel):
         auto_expire_configuration: main_models.CreateKyuubiTokenRequestAutoExpireConfiguration = None,
         member_arns: List[str] = None,
         name: str = None,
+        spark_role: List[str] = None,
         token: str = None,
         region_id: str = None,
     ):
         self.auto_expire_configuration = auto_expire_configuration
         self.member_arns = member_arns
         self.name = name
+        self.spark_role = spark_role
         self.token = token
         self.region_id = region_id
 
@@ -40,6 +42,9 @@ class CreateKyuubiTokenRequest(DaraModel):
         if self.name is not None:
             result['name'] = self.name
 
+        if self.spark_role is not None:
+            result['sparkRole'] = self.spark_role
+
         if self.token is not None:
             result['token'] = self.token
 
@@ -59,6 +64,9 @@ class CreateKyuubiTokenRequest(DaraModel):
 
         if m.get('name') is not None:
             self.name = m.get('name')
+
+        if m.get('sparkRole') is not None:
+            self.spark_role = m.get('sparkRole')
 
         if m.get('token') is not None:
             self.token = m.get('token')